pyproject.toml
[project]
name = "homework-bot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "B_K", email = "chayanongg2550@gmail.com" }
]
dependencies = [
    "httpx>=0.25.0",
    "py-cord[speed]>=2.4.1",
    "python-dotenv>=1.0.0",
    "databases[sqlite]>=0.8.0",
    "cryptography>=41.0.4",
    "seaborn>=0.13.0",
    "pycord-multicog>=1.1.0",
    "cacheout>=0.15.0",
    "psutil>=5.9.7",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
dev-dependencies = []
